Add enhanced gf mappings
A module with relative path "foo/bar" is searched for by jq in
"foo/bar.jq" and "foo/bar/bar.jq" in the given search paths (either
provided by the -L option, or by default in "/.jq" and
"$ORIGIN/../lib").

While [i, [d, :checkpath! (and friends) always use includeexpr first,
gf uses includeexpr as a fallback only when "foo/bar" can't be found
in Vim's path (as file or directory). This means, pressing gf on
"foo/bar" edits the directory "foo/bar" even though we want to go to the
file "foo/bar/bar.jq". The following gf mappings solve this issue.

Note: currently no error is emitted when gf is pressed while the cursor
is on a character that does not satisfy isfname.