Work around 'contains' parsing in syntaxcomplete
In order to find list of completions syntaxcomplete plugin parses output
of `:syntax list`. This output contains `:syntax` command's keywords and
syntaxcomplete does some sloppy parsing for them:

https://github.com/vim/vim/blob/a14e6975478adeddcc2161edc1ec611016aa89f3/runtime/autoload/syntaxcomplete.vim#L554-L567

Because of this logic everything after 'contains' on the line is
ignored and doesn't show up in completions.

'contains' is a valid jq keyword and currently appears somewhere in the
middle of the output for `jqBuiltinFunction` group. This causes rest of
the keywords in the group to be ignored for syntaxcomplete's completion.

Fix by moving 'contains' to a separate syntax group where it will always
be a single keyword. It will still be missing from the completions list
itself, but at least it won't break syntaxcomplete's parsing for other
keywords.